backend/requirements.txt
fastapi
uvicorn[standard]
pypdf
chromadb
pydantic-settings
python-multipart
Pillow
huggingface-hub
ultralytics
openai
